requirements.txt
# Python 3.10+ required
# Install with: pip install -r requirements.txt

# PyTorch with CUDA 11.8
--extra-index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# HuggingFace ecosystem
transformers>=4.36.0
datasets>=2.15.0
accelerate>=0.25.0
safetensors>=0.4.0
huggingface_hub>=0.19.0

# MLflow for experiment tracking
mlflow>=2.9.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Utilities
numpy>=1.24.0
einops>=0.7.0
tqdm>=4.66.0
pyyaml>=6.0.0
omegaconf>=2.3.0

# Metrics
torchmetrics>=1.2.0
torch-fidelity>=0.3.0

# Visualization
matplotlib>=3.8.0
tensorboard>=2.15.0

# Notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0
